Story Model: moved into own dir with subdirs for instance/static methods
