+++ models/index.js
@@ -1,6 +1,6 @@
 const { Clap, MAX_CLAP_COUNT } = require('./clap');
 const { User, SALT_ROUNDS } = require('./user');
-const Story = require('./story');
+const { Story } = require('./story');
 
 module.exports = {
   User,
