+++ models/story/index.js
@@ -1,0 +1,9 @@
+const Story = require('./story');
+const instanceMethods = require('./story-instance-methods');
+const staticMethods = require('./story-static-methods');
+
+module.exports = {
+  Story,
+  staticMethods,
+  instanceMethods,
+};
